Unify v1 and v2 DESCRIBE TABLE
